requirements.txt
pyyaml==5.4.1
numpy==1.19.5
pandas==1.3.3
tabulate==0.8.9
xgboost==1.4.2
scikit-optimize==0.9.0
scikit-learn==0.24.2
matplotlib==3.4.3
seaborn==0.11.2
requests==2.26.0
beautifulsoup4==4.10.0
optuna==2.10.1